Worked on task 7: Modified YOLOv3.h and added UML folder

+++ include/YOLOv3.h
@@ -34,7 +34,7 @@ class YOLOv3 {
   /**
    * @brief private variable for storing inputHeight.
    */
-  double inputHeight;
+  int inputHeight;
 
  public:
   /**
